Функционал группировки списка ошибок выделен в отдельный сервис
